Move #[cfg] functions to sys
