+++ src/shell/job_control.rs
@@ -1,13 +1,15 @@
-#[cfg(all(unix, not(target_os = "redox")))] use libc::{self, pid_t};
 use std::fmt;
 use std::thread::{sleep, spawn};
 use std::time::Duration;
 use std::sync::{Arc, Mutex};
-use super::foreground::{ForegroundSignals, BackgroundResult};
+use super::foreground::BackgroundResult;
 use super::signals;
 use super::status::*;
 use super::Shell;
 use sys;
+
+pub use sys::job_control::watch_background;
+use sys::job_control as self_sys;
 
 /// When given a process ID, that process's group will be assigned as the foreground process group.
 pub fn set_foreground_as(pid: u32) {
@@ -26,14 +28,16 @@ pub trait JobControl {
     fn handle_signal(&self, signal: i32) -> bool;
     fn foreground_send(&self, signal: i32);
     fn background_send(&self, signal: i32);
-    fn watch_foreground <F, D> (
+    fn watch_foreground<F, D>(
         &mut self,
         pid: u32,
         last_pid: u32,
         get_command: F,
-        drop_command: D
-        ) -> i32 where F: FnOnce() -> String,
-                       D: FnMut(i32);
+        drop_command: D,
+    ) -> i32
+    where
+        F: FnOnce() -> String,
+        D: FnMut(i32);
     fn send_to_background(&mut self, child: u32, state: ProcessState, command: String);
 }
 
@@ -42,7 +46,7 @@ pub trait JobControl {
 pub enum ProcessState {
     Running,
     Stopped,
-    Empty
+    Empty,
 }
 
 impl fmt::Display for ProcessState {
@@ -50,103 +54,38 @@ impl fmt::Display for ProcessState {
         match *self {
             ProcessState::Running => write!(f, "Running"),
             ProcessState::Stopped => write!(f, "Stopped"),
-            ProcessState::Empty   => write!(f, "Empty"),
+            ProcessState::Empty => write!(f, "Empty"),
         }
     }
 }
 
-#[cfg(target_os = "redox")]
-pub fn watch_background (
-    fg: Arc<ForegroundSignals>,
-    processes: Arc<Mutex<Vec<BackgroundProcess>>>,
-    pid: u32,
-    njob: usize
-) {
-    // TODO: Implement this using syscall::call::waitpid
-}
-
-#[cfg(all(unix, not(target_os = "redox")))]
-pub fn watch_background (
-    fg: Arc<ForegroundSignals>,
-    processes: Arc<Mutex<Vec<BackgroundProcess>>>,
-    pid: u32,
-    njob: usize
-) {
-    use nix::sys::wait::*;
-    let mut fg_was_grabbed = false;
-    loop {
-        if !fg_was_grabbed {
-            if fg.was_grabbed(pid) { fg_was_grabbed = true; }
-        }
-        match waitpid(-(pid as pid_t), Some(WUNTRACED | WCONTINUED | WNOHANG)) {
-            Ok(WaitStatus::Exited(_, status)) => {
-                if !fg_was_grabbed {
-                    eprintln!("ion: ([{}] {}) exited with {}", njob, pid, status);
-                }
-                let mut processes = processes.lock().unwrap();
-                let process = &mut processes.iter_mut().nth(njob).unwrap();
-                process.state = ProcessState::Empty;
-                if fg_was_grabbed { fg.reply_with(status); }
-                break
-            },
-            Ok(WaitStatus::Stopped(pid, _)) => {
-                if !fg_was_grabbed {
-                    eprintln!("ion: ([{}] {}) Stopped", njob, pid);
-                }
-                let mut processes = processes.lock().unwrap();
-                let process = &mut processes.iter_mut().nth(njob).unwrap();
-                if fg_was_grabbed {
-                    fg.reply_with(TERMINATED as i8);
-                    fg_was_grabbed = false;
-                }
-                process.state = ProcessState::Stopped;
-            },
-            Ok(WaitStatus::Continued(pid)) => {
-                if !fg_was_grabbed {
-                    eprintln!("ion: ([{}] {}) Running", njob, pid);
-                }
-                let mut processes = processes.lock().unwrap();
-                let process = &mut processes.iter_mut().nth(njob).unwrap();
-                process.state = ProcessState::Running;
-            },
-            Ok(_) => (),
-            Err(why) => {
-                eprintln!("ion: ([{}] {}) errored: {}", njob, pid, why);
-                let mut processes = processes.lock().unwrap();
-                let process = &mut processes.iter_mut().nth(njob).unwrap();
-                process.state = ProcessState::Empty;
-                if fg_was_grabbed { fg.errored(); }
-                break
-            }
-        }
-        sleep(Duration::from_millis(100));
-    }
-}
-
-pub fn add_to_background (
+pub fn add_to_background(
     processes: Arc<Mutex<Vec<BackgroundProcess>>>,
     pid: u32,
     state: ProcessState,
-    command: String
+    command: String,
 ) -> u32 {
     let mut processes = processes.lock().unwrap();
-    match (*processes).iter().position(|x| x.state == ProcessState::Empty) {
+    match (*processes)
+        .iter()
+        .position(|x| x.state == ProcessState::Empty)
+    {
         Some(id) => {
             (*processes)[id] = BackgroundProcess {
                 pid: pid,
                 ignore_sighup: false,
                 state: state,
-                name: command
+                name: command,
             };
             id as u32
-        },
+        }
         None => {
             let njobs = (*processes).len();
             (*processes).push(BackgroundProcess {
                 pid: pid,
                 ignore_sighup: false,
                 state: state,
-                name: command
+                name: command,
             });
             njobs as u32
         }
@@ -162,13 +101,15 @@ pub struct BackgroundProcess {
     pub pid: u32,
     pub ignore_sighup: bool,
     pub state: ProcessState,
-    pub name: String
+    pub name: String,
 }
 
 impl<'a> JobControl for Shell<'a> {
     fn set_bg_task_in_foreground(&self, pid: u32, cont: bool) -> i32 {
         // Resume the background task, if needed.
-        if cont { signals::resume(pid); }
+        if cont {
+            signals::resume(pid);
+        }
         // Pass the TTY to the background job
         set_foreground_as(pid);
         // Signal the background thread that is waiting on this process to stop waiting.
@@ -180,7 +121,7 @@ impl<'a> JobControl for Shell<'a> {
             match self.foreground_signals.was_processed() {
                 Some(BackgroundResult::Status(stat)) => break stat as i32,
                 Some(BackgroundResult::Errored) => break TERMINATED,
-                None => sleep(Duration::from_millis(25))
+                None => sleep(Duration::from_millis(25)),
             }
         };
         // Have the shell reclaim the TTY
@@ -197,106 +138,30 @@ impl<'a> JobControl for Shell<'a> {
                     while let Some(signal) = self.next_signal() {
                         if signal != sys::SIGTSTP {
                             self.background_send(signal);
-                            break 'event
+                            break 'event;
                         }
                     }
                     sleep(Duration::from_millis(100));
-                    continue 'event
+                    continue 'event;
                 }
             }
-            return
+            return;
         }
         self.exit(TERMINATED);
     }
 
-    #[cfg(all(unix, not(target_os = "redox")))]
-    fn watch_foreground <F: FnOnce() -> String, D: FnMut(i32)> (
+    fn watch_foreground<F, D>(
         &mut self,
         pid: u32,
         last_pid: u32,
         get_command: F,
-        mut drop_command: D,
-    ) -> i32 {
-        use nix::sys::wait::{waitpid, WaitStatus, WUNTRACED};
-        use nix::sys::signal::Signal;
-        use nix::{Error, Errno};
-        let mut exit_status = 0;
-        loop {
-            match waitpid(-(pid as pid_t), Some(WUNTRACED)) {
-                Ok(WaitStatus::Exited(pid, status)) => {
-                    if pid == (last_pid as i32) {
-                        break status as i32
-                    } else {
-                        drop_command(pid);
-                        exit_status = status;
-                    }
-                }
-                Ok(WaitStatus::Signaled(_, signal, _)) => {
-                    eprintln!("ion: process ended by signal");
-                    if signal == Signal::SIGTERM {
-                        self.handle_signal(libc::SIGTERM);
-                        self.exit(TERMINATED);
-                    } else if signal == Signal::SIGHUP {
-                        self.handle_signal(libc::SIGHUP);
-                        self.exit(TERMINATED);
-                    } else if signal == Signal::SIGINT {
-                        self.foreground_send(libc::SIGINT as i32);
-                        self.break_flow = true;
-                    }
-                    break TERMINATED;
-                },
-                Ok(WaitStatus::Stopped(pid, _)) => {
-                    self.send_to_background(pid as u32, ProcessState::Stopped, get_command());
-                    self.break_flow = true;
-                    break TERMINATED
-                },
-                Ok(_) => (),
-                // ECHILD signifies that all children have exited
-                Err(Error::Sys(Errno::ECHILD)) => break exit_status as i32,
-                Err(why) => {
-                    eprintln!("ion: process doesn't exist: {}", why);
-                    break FAILURE
-                }
-            }
-        }
-    }
-
-    #[cfg(target_os = "redox")]
-    fn watch_foreground <F: FnOnce() -> String, D: FnMut(i32)> (
-        &mut self,
-        pid: u32,
-        _last_pid: u32,
-        _get_command: F,
-        mut drop_command: D,
-    ) -> i32 {
-        use std::io::{self, Write};
-        use std::os::unix::process::ExitStatusExt;
-        use std::process::ExitStatus;
-        use syscall;
-
-        loop {
-            let mut status_raw = 0;
-            match syscall::waitpid(pid as usize, &mut status_raw, 0) {
-                Ok(0) => (),
-                Ok(_pid) => {
-                    let status = ExitStatus::from_raw(status_raw as i32);
-                    if let Some(code) = status.code() {
-                        break code
-                    } else {
-                        let stderr = io::stderr();
-                        let mut stderr = stderr.lock();
-                        let _ = stderr.write_all(b"ion: child ended by signal\n");
-                        break TERMINATED
-                    }
-                },
-                Err(err) => {
-                    let stderr = io::stderr();
-                    let mut stderr = stderr.lock();
-                    let _ = writeln!(stderr, "ion: failed to wait: {}", err);
-                    break 100 // TODO what should we return here?
-                }
-            }
-        }
+        drop_command: D,
+    ) -> i32
+    where
+        F: FnOnce() -> String,
+        D: FnMut(i32),
+    {
+        self_sys::watch_foreground(self, pid, last_pid, get_command, drop_command)
     }
 
     /// Send a kill signal to all running foreground tasks.
@@ -349,6 +214,8 @@ impl<'a> JobControl for Shell<'a> {
         if signal == sys::SIGTERM || signal == sys::SIGHUP {
             self.background_send(signal);
             true
-        } else { false }
+        } else {
+            false
+        }
     }
 }
