+++ src/shell/variables.rs
@@ -14,6 +14,8 @@ use types::{
     Array,
 };
 
+use ::sys::variables as self_sys;
+
 #[derive(Debug)]
 pub struct Variables {
     pub arrays:    ArrayVariableContext,
@@ -196,7 +198,7 @@ impl Variables {
                         }
                     }
                     Err(_) => {
-                        if let Some(home) = get_user_home(tilde_prefix) {
+                        if let Some(home) = self_sys::get_user_home(tilde_prefix) {
                             return Some(home + remainder);
                         }
                     }
@@ -221,29 +223,6 @@ impl Variables {
 
         None
     }
-}
-
-#[cfg(all(unix, not(target_os = "redox")))]
-fn get_user_home(username: &str) -> Option<String> {
-    use users_unix::get_user_by_name;
-    use users_unix::os::unix::UserExt;
-
-    match get_user_by_name(username) {
-        Some(user) => Some(user.home_dir().to_string_lossy().into_owned()),
-        None => None,
-    }
-}
-
-#[cfg(target_os = "redox")]
-fn get_user_home(_username: &str) -> Option<String> {
-    // TODO
-    None
-}
-
-#[cfg(not(any(unix, target_os = "redox")))]
-fn get_user_home(_username: &str) -> Option<String> {
-    // TODO
-    None
 }
 
 #[cfg(test)]
